g2 hack fest/requirements.txt
Flask==2.3.3
firebase-admin==6.2.0
python-dotenv==1.0.0
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
google-cloud-firestore==2.13.1 